refactor: rename config file format
