+++ src/config/authors-config.ts
