+++ src/config/context-config.ts
@@ -1,0 +1,4 @@
+export const DOCUMENTS = {
+  MAX_NUMBER_OF_DOCUMENTS: 6,
+  MAX_LENGTH_PER_DOCUMENT: 2000,
+}
